Fix circular import in tests
